requirements.txt
anyascii==0.3.1
asgiref==3.6.0
beautifulsoup4==4.11.1
bleach==4.1.0
captcha==0.4
certifi==2022.12.7
charset-normalizer==2.1.1
crispy-bootstrap5==0.7
Django==4.1.5
django-cogwheels==0.3
django-comments-xtd==2.9.9
django-contrib-comments==2.2.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-mjml==1.0
django-modelcluster==6.0
django-permissionedforms==0.1
django-recaptcha==3.0.0
django-social-share==2.3.0
django-taggit==3.1.0
django-treebeard==4.6.0
djangorestframework==3.14.0
docutils==0.19
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
l18n==2021.3
Markdown==3.4.1
mysqlclient==2.1.1
openpyxl==3.0.10
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pytz==2022.7
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
telepath==0.3
urllib3==1.26.13
wagtail==4.1.1
wagtail-birdsong==1.1.0
wagtail-django-recaptcha==1.0
wagtail-markdown==0.10.0
wagtailmenus==3.1.3
webencodings==0.5.1
whitenoise==6.4.0
Willow==1.4.1
